refactor: split visit page actions into domain modules
Extract 17 action handlers from the 805-line +page.server.ts into
5 focused modules: billing, items, clinical notes, files, and
hospitalisation. Main file reduced to 248 lines.

Closes #433
