+++ src/routes/(app)/visit/[id]/+page.server.ts
@@ -3,12 +3,10 @@ import type { Actions, PageServerLoad } from './$types';
 import type {
 	MedicalActsResponse,
 	SurgicalActsResponse,
-	BillsMethodOptions,
 	VisitsResponse,
 	CagesResponse,
 	InventoryItemResponse,
 	Visit,
-	ItemMetadata,
 	DoctorsResponse,
 	ToilettageResponse
 } from '$types';
@@ -26,16 +24,31 @@ import {
 	updateVisitTreatmentSchema
 } from '$lib/schemas/visit';
 import type { RecordModel } from 'pocketbase';
-import { setError, superValidate, withFiles } from 'sveltekit-superforms/server';
+import { setError, superValidate } from 'sveltekit-superforms/server';
 import BillService from '$lib/services/bill';
 import { zod } from 'sveltekit-superforms/adapters';
 import { removeSchema } from '$lib/schemas';
-import { InventoryService } from '$lib/services/inventory';
 import { cageCompare } from '$utils/cage';
 import { unknownClient } from '$utils/client';
 import { unknownAnimal } from '$lib/utils/animal';
 import { updateHospitCompletedStateSchema } from '$lib/schemas/hospit';
 import { buildFileProxyUrl } from '$utils/file';
+
+import { addPayment } from './visit-billing-actions';
+import {
+	addToilettage,
+	removeToilettage,
+	addMedicalActs,
+	removeMedicalAct,
+	addSurgicalActs,
+	removeSurgicalAct,
+	addVisitStoreItem,
+	removeVisitStoreItem,
+	updateVisitItem
+} from './visit-item-actions';
+import { updateDiagnostic, updateActions, updateTreatment } from './visit-clinical-actions';
+import { addFile, removeFile } from './visit-file-actions';
+import { updateHospit, removeHospit } from './visit-hospit-actions';
 
 export const load = (async ({ params, locals: { pb }, url: { searchParams } }) => {
 	const { id } = params;
@@ -215,591 +228,21 @@ export const actions = {
 		}
 	},
 
-	addPayment: async ({ locals: { pb, user }, request }) => {
-		const form = await superValidate(request, zod(payVisitSchema), { id: 'pay-visit' });
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { amount, id, incash, method, outcash, description: desc } = form.data;
-			const item = await pb.collection('visits').getOne<VisitsResponse>(id);
-
-			if (!item) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			const billService = new BillService(pb, item);
-			const expandedVisit = await billService.getClientExpandedVisit();
-
-			const visitLink = `visite client: <a href="/visit/${item.id}" class="text-blue-500">${expandedVisit.animal.client.name}</a>`;
-
-			const description = desc.length ? `${desc}\n ${visitLink}` : visitLink;
-
-			await billService.pay(
-				amount,
-				method as BillsMethodOptions,
-				incash,
-				outcash,
-				description,
-				user?.id ?? ''
-			);
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	addToilettage: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(addVisitItemsSchema), { id: 'add-toilettage' });
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, items } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'toilettage+': items
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	removeToilettage: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(removeVisitItemSchema), {
-			id: 'remove-toilettage'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, item } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'toilettage-': item
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	addFile: async ({ locals: { pb }, request }) => {
-		const formData = await request.formData();
-		const form = await superValidate(formData, zod(addVisitFileSchema), { id: 'add-file' });
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, form.data);
-
-			return withFiles({ form });
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, "Échec de l'ajout du fichier", { status: 500 });
-		}
-	},
-
-	removeFile: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(removeVisitFileSchema), { id: 'remove-file' });
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, file } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'files-': file
-			});
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	updateDiagnostic: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(updateVisitDiagnosticSchema), {
-			id: 'update-diagnostic'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, observations } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				observations
-			});
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	updateActions: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(updateVisitActionsSchema), {
-			id: 'update-actions'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, actions } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				actions
-			});
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	addVisitStoreItem: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(addVisitItemsSchema), {
-			id: 'add-visit-store-item'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, items } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-			const inventoryService = new InventoryService(pb);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'store_items+': items
-			});
-
-			await Promise.all(items.map((itemId) => inventoryService.decreaseItemQuantity(itemId, 1)));
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	removeVisitStoreItem: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(removeVisitItemSchema), {
-			id: 'remove-visit-store-item'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, item } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse<ItemMetadata[]>>(id);
-			const billService = new BillService(pb, visit);
-			const inventoryService = new InventoryService(pb);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			const itemQuantity =
-				visit.item_metadata?.find((metdata) => metdata.item === item)?.quantity ?? 1;
-
-			await inventoryService.increaseItemQuantity(item, itemQuantity);
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'store_items-': item
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	addMedicalActs: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(addVisitItemsSchema), { id: 'add-medical-acts' });
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, items } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Visite non trouvée', { status: 500 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'medical_acts+': items
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	removeMedicalAct: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(removeVisitItemSchema), {
-			id: 'remove-medical-act'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, item } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'medical_acts-': item
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	addSurgicalActs: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(addVisitItemsSchema), {
-			id: 'add-surgical-acts'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, items } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'surgical_acts+': items
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	removeSurgicalAct: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(removeVisitItemSchema), {
-			id: 'remove-surgical-act'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, item } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				'surgical_acts-': item
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	updateHospit: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(updateVisitHospitalisationSchema), {
-			id: 'update-hospit'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, cage, end, start, note, treatment, price } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			if (visit.hospit) {
-				await pb.collection('hospitalisation').update(visit.hospit, {
-					cage,
-					end,
-					start,
-					note,
-					treatment,
-					price
-				});
-			} else {
-				const hospit = await pb.collection('hospitalisation').create({
-					cage,
-					end,
-					start,
-					note,
-					treatment,
-					visit: visit.id,
-					price
-				});
-
-				await pb.collection('visits').update(id, {
-					...visit,
-					hospit: hospit.id
-				});
-			}
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	removeHospit: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(removeSchema), { id: 'remove-hospit' });
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-			const billService = new BillService(pb, visit);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			if (visit.hospit) {
-				await pb.collection('hospitalisation').delete(visit.hospit);
-				await pb.collection('visits').update<VisitsResponse>(id, { ...visit, hospit: null });
-				await billService.update();
-			}
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	updateTreatment: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(updateVisitTreatmentSchema), {
-			id: 'update-treatment'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, treatment } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse>(id);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				treatment
-			});
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	},
-
-	updateVisitItem: async ({ locals: { pb }, request }) => {
-		const form = await superValidate(request, zod(updateVisitItemSchema), {
-			id: 'update-visit-item'
-		});
-
-		try {
-			if (!form.valid) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-
-			const { id, discount, quantity, item, type } = form.data;
-			const visit = await pb.collection('visits').getOne<VisitsResponse<ItemMetadata[]>>(id);
-
-			if (!visit) {
-				return setError(form, 'Données invalides', { status: 400 });
-			}
-			const billService = new BillService(pb, visit);
-			const inventoryService = new InventoryService(pb);
-			const items: ItemMetadata[] = visit.item_metadata || [];
-			const existingIndex = items.findIndex((metadata) => metadata.item === item);
-
-			if (type === 'inventory_item') {
-				const currentQuantity = existingIndex !== -1 ? items[existingIndex].quantity : 1;
-				if (currentQuantity > quantity) {
-					await inventoryService.increaseItemQuantity(item, currentQuantity - quantity);
-				} else if (currentQuantity < quantity) {
-					await inventoryService.decreaseItemQuantity(item, quantity - currentQuantity);
-				}
-			}
-
-			if (existingIndex !== -1) {
-				items[existingIndex].quantity = quantity;
-				items[existingIndex].discount = discount;
-			} else {
-				items.push({ item, discount, quantity });
-			}
-
-			await pb.collection('visits').update(id, {
-				...visit,
-				item_metadata: JSON.stringify(items)
-			});
-
-			await billService.update();
-
-			return { form };
-		} catch (error) {
-			console.error(error);
-
-			return setError(form, 'Échec de la mise à jour de la visite', { status: 500 });
-		}
-	}
+	addPayment,
+	addToilettage,
+	removeToilettage,
+	addMedicalActs,
+	removeMedicalAct,
+	addSurgicalActs,
+	removeSurgicalAct,
+	addVisitStoreItem,
+	removeVisitStoreItem,
+	updateVisitItem,
+	updateDiagnostic,
+	updateActions,
+	updateTreatment,
+	addFile,
+	removeFile,
+	updateHospit,
+	removeHospit
 } satisfies Actions;
